requirements.txt
numpy>=1.17.0
pytorch>=2.0.1